feat: rename `hclfmt` command to `hcl format`
